+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Improve survey models for BESSI [#40](https://github.com/rokwire/polls-building-block/issues/40)
+- Send notifications for group polls directly through Groups BB [#30](https://github.com/rokwire/polls-building-block/issues/30)
+- Set appID and orgID for notifications [#26](https://github.com/rokwire/polls-building-block/issues/26)
 
 ## [1.3.0] - 2022-11-10
 ### Added
@@ -22,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created CRUD APIs for Survey Responses [#24](https://github.com/rokwire/polls-building-block/issues/24)
 
 ## Fixed
- - Fix docs [#21](https://github.com/rokwire/polls-building-block/issues/21)
+- Fix docs [#21](https://github.com/rokwire/polls-building-block/issues/21)
 
 ## [1.0.21] - 2022-08-11
 ### Fixed